CFY-5926 Storage manager refactor
- Got rid of the file storage manager - sqlite will be used in its stead with the default sql storage manager
- Got rid of ES storage manager
- Got rid of unnecessary fluff in the blueprints manager
