+++ rest-service/manager_rest/test/endpoints/test_events_v2.py
@@ -13,7 +13,6 @@
 #  * limitations under the License.
 
 from nose.plugins.attrib import attr
-from mock import patch
 
 from manager_rest.test import base_test
 from manager_rest.resources_v2 import Events
@@ -42,18 +41,16 @@ class EventsTest(base_test.BaseServerTestCase):
     @attr(client_min_version=2.1,
           client_max_version=base_test.LATEST_API_VERSION)
     def test_delete_events(self):
-        def delete_events(bpm_self, events_list):
-            for index, event in enumerate(events_list):
-                self.assertEqual(event['doc_type'], 'cloudify_log')
-                self.assertEqual(event['id'], 'id_{0}'.format(index + 1))
-                self.assertEqual(event['index'], 'logstash-1')
+        def delete_events(events_list):
+            for index, event in enumerate(events_list['hits']['hits']):
+                self.assertEqual(event['_source']['type'], 'cloudify_log')
+                self.assertEqual(event['_id'], 'id_{0}'.format(index + 1))
+                self.assertEqual(event['_index'], 'logstash-1')
 
         ManagerElasticsearch.search_events = self._mock_es_search_delete
-        patch_path = ('manager_rest.blueprints_manager.BlueprintsManager.'
-                      'delete_events')
-        with patch(patch_path, delete_events):
-            response = self.client.events.delete('dep_id', include_logs=True)
-            self.assertEqual(response.items, [5])
+        ManagerElasticsearch.delete_events = staticmethod(delete_events)
+        response = self.client.events.delete('dep_id', include_logs=True)
+        self.assertEqual(response.items, [5])
 
     def test_build_query(self):
         self.maxDiff = None
