+++ rest-service/manager_rest/test/endpoints/test_storage_manager.py
@@ -25,7 +25,7 @@ class StorageManagerTests(base_test.BaseServerTestCase):
 
     def test_store_load_delete_blueprint(self):
         now = utils.get_formatted_timestamp()
-        sm = storage_manager._get_instance()
+        sm = storage_manager.get_storage_manager()
         blueprint = models.BlueprintState(id='blueprint-id',
                                           created_at=now,
                                           updated_at=now,
@@ -33,7 +33,7 @@ class StorageManagerTests(base_test.BaseServerTestCase):
                                           plan={'name': 'my-bp'},
                                           source='bp-source',
                                           main_file_name='aaa')
-        sm.put_blueprint('blueprint-id', blueprint)
+        sm.put_blueprint(blueprint)
         blueprint_from_list = sm.list_blueprints().items[0]
         blueprint_restored = sm.get_blueprint('blueprint-id')
         bp_from_delete = sm.delete_blueprint('blueprint-id')
@@ -47,7 +47,7 @@ class StorageManagerTests(base_test.BaseServerTestCase):
 
     def test_get_blueprint_deployments(self):
         now = utils.get_formatted_timestamp()
-        sm = storage_manager._get_instance()
+        sm = storage_manager.get_storage_manager()
         blueprint = models.BlueprintState(id='blueprint-id',
                                           created_at=now,
                                           updated_at=now,
@@ -55,7 +55,7 @@ class StorageManagerTests(base_test.BaseServerTestCase):
                                           plan={'name': 'my-bp'},
                                           source='bp-source',
                                           main_file_name='aaa')
-        sm.put_blueprint('blueprint-id', blueprint)
+        sm.put_blueprint(blueprint)
 
         deployment1 = models.Deployment(id='dep-1',
                                         created_at=now,
@@ -71,7 +71,7 @@ class StorageManagerTests(base_test.BaseServerTestCase):
                                         groups={},
                                         scaling_groups={},
                                         outputs={})
-        sm.put_deployment('dep-1', deployment1)
+        sm.put_deployment(deployment1)
 
         deployment2 = models.Deployment(id='dep-2',
                                         created_at=now,
@@ -87,7 +87,7 @@ class StorageManagerTests(base_test.BaseServerTestCase):
                                         groups={},
                                         scaling_groups={},
                                         outputs={})
-        sm.put_deployment('dep-2', deployment2)
+        sm.put_deployment(deployment2)
 
         deployment3 = models.Deployment(id='dep-3',
                                         created_at=now,
@@ -103,7 +103,7 @@ class StorageManagerTests(base_test.BaseServerTestCase):
                                         groups={},
                                         scaling_groups={},
                                         outputs={})
-        sm.put_deployment('dep-3', deployment3)
+        sm.put_deployment(deployment3)
 
         blueprint_deployments = \
             sm.list_blueprint_deployments('blueprint-id').items
@@ -159,13 +159,13 @@ class StorageManagerTests(base_test.BaseServerTestCase):
                                           plan={'name': 'my-bp'},
                                           source='bp-source',
                                           main_file_name='aaa')
-        sm = storage_manager._get_instance()
-        sm.put_blueprint('blueprint-id', blueprint)
+        sm = storage_manager.get_storage_manager()
+        sm.put_blueprint(blueprint)
 
         blueprint_restored = sm.get_blueprint('blueprint-id',
                                               {'id', 'created_at'})
         self.assertEquals('blueprint-id', blueprint_restored.id)
         self.assertEquals(now, blueprint_restored.created_at)
-        self.assertEquals(None, blueprint_restored.updated_at)
-        self.assertEquals(None, blueprint_restored.plan)
-        self.assertEquals(None, blueprint_restored.main_file_name)
+        self.assertFalse(hasattr(blueprint_restored, 'updated_at'))
+        self.assertFalse(hasattr(blueprint_restored, 'plan'))
+        self.assertFalse(hasattr(blueprint_restored, 'main_file_name'))
